+++ lib/src/core/model/api_request/get_reaction_request.dart
@@ -39,4 +39,9 @@ class GetReactionRequest {
         "reactionName": reactionName,
         "options": options?.toJson(),
       }..removeWhere((key, value) => value == null);
+
+  int getHashCode() {
+    final jsonString = (toJson()..remove('options')).toString();
+    return jsonString.hashCode;
+  }
 }